meta/Parametrization.m: parametrization of indexed couplings
To use Newton's method, one needs to differentiate constraints with
respect to the Wilson coefficients.  Therefore, a single parameter
that one refers to in multiple different ways must be identified:
e.g. the real parts of mq2[1,2] and mq2[2,1] are really identical
parameters.  The added functions are intended for parametrizing an
indexed coupling in terms of a set of independent real variables.  For
instance, a 2 X 2 hermitian matrix is parametrized as

{{Re[mq2[1,1]], -I Im[mq2[2,1]] + Re[mq2[2,1]]},
 { I Im[mq2[2,1]] + Re[mq2[2,1]], Re[mq2[2,2]]}}