meta/TextFormatting.m: add WrapText[], a faster text wrapper
WrapText[text_String, maxWidth_Integer, indentation_Integer] is much
faster than WrapLines[], and therefore is better suited to the lattice
method which can produce a long C source.  It does not split
multi-character operators such as ++, <<, *=.  It takes care of string
constants containing \".  The semantics of maxWidth is different from
that for WrapLines[]: WrapText[] allows a line (excluding "\n") to be
maxWidth characters long.