Check full Lorentz structure to find a mixing matrix
In the previous commit, RewriteUnrotatedField[] determined the fermion
mixing matrix based on the following two factors:

  (a) whether the coupling has the form Cp[__][PL] or Cp[__][PR]
  (b) whether the vertex type is FFS or FFV

Now, it examines instead the full Lorentz structure returned by
SARAH`Vertex.  Each combination of the above two factors reduces to:

  PL, PR, LorentzProduct[gamma[_], PL], LorentzProduct[gamma[_], PR],

The background reason for this change was that I was not sure whether
FFV might denote also a (non-renormalizable) dipole type operator or
not.