Introduce fflite module that compute loop functions
The fflite module is a subset of the FF library that is adapted to and
distributed as part of LoopTools 2.8.  The purpose to extract this
subset from LoopTools was to have thread-safe loop function routines
that accept complex (and therefore also negative) mass^2 arguments.
Calls to fflite subroutines bypass the internal cache of LoopTools.
It is unclear whether this would lead to a performance gain or loss.